Documentation: Adds README.md for the project.
README.md modified to provide information that will allow others
to contribute to the project. Adds and renames some npm scripts to
reduce the extent to which contributors need to have knowledge of
the third party libraries involved. This also has the added
benefit of being able to substitute third party libraries without
having to modify build scripts.